% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grazing_soc.R
\name{grazing_soc}
\alias{grazing_soc}
\title{Equations for SNAP model}
\usage{
grazing_soc(x)
}
\arguments{
\item{x}{(data frame) Data input containing values for each of the associated parameters}
}
\value{
list of key variables
}
\description{
Equations for SNAP model
}
\examples{
\dontrun{
article_pdf_download(infilepath = "/data/isi_searches", outfilepath = "data")
}
}
