% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grazing_soc_data.R
\name{grazing_soc_data}
\alias{grazing_soc_data}
\title{Equations for SNAP model when calculating with data}
\usage{
grazing_soc_data(x)
}
\arguments{
\item{x}{(data frame) Data input containing values for each of the associated parameters}
}
\value{
list of key variables
}
\description{
Equations for SNAP model when calculating with data
}
