% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grazing_soc_noprice.R
\name{grazing_soc_noprice}
\alias{grazing_soc_noprice}
\title{Equations for SNAP model, without cost and revenue calculations}
\usage{
grazing_soc_noprice(x)
}
\arguments{
\item{x}{(data frame) Data input containing values for each of the associated parameters}
}
\value{
list of key variables
}
\description{
Equations for SNAP model, without cost and revenue calculations
}
