% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_data.R
\name{samp_data}
\alias{samp_data}
\title{Resample data}
\usage{
samp_data(data, n = 10000)
}
\arguments{
\item{data}{Data frame to be resampled}

\item{n}{(numeric) Number of samples to draw for resampling procedure}
}
\value{
Data frame to pass to the data argumen in snap()
}
\description{
Resample data
}
\examples{
{
data <- samp_data(fake_data, n = 100)
snap(data = data)
}
}
