% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap.R
\name{snap}
\alias{snap}
\title{Run the SNAP model}
\usage{
snap(data = NA, input = NA, nsamp = 10000, plot_vars = NA)
}
\arguments{
\item{data}{Data frame containing individual observations that can be used to calculate soil C for each observation}

\item{input}{An alternative data source to data. This provides a lower and upper bound, and distribution type, that are sampled from in a bootstrapping procedure}

\item{nsamp}{(numeric) Number of samples to draw for bootstrapping procedure}

\item{plot_vars}{Which variables to include in plot of variable importance}
}
\value{
Output values for multiple variables
}
\description{
Run the SNAP model
}
\examples{
{
cost <- snap(
  input = input_table, nsamp = 1000, plot_vars = c("carbon_project_NPV")
)
head(cost$y)
}
}
