% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_data.R
\name{snap_data}
\alias{snap_data}
\title{Run SNAP model}
\usage{
snap_data(input = NA)
}
\arguments{
\item{input}{Input data to use}

\item{data}{(logical) Binary of whether data are being used}

\item{nsamp}{(numeric) Number of samples to draw for bootstrapping procedure}

\item{plot_vars}{Which variables to include in plot of variable importance}
}
\value{
Output values for multiple variables
}
\description{
Run SNAP model
}
\examples{
\dontrun{
article_pdf_download(infilepath = "/data/isi_searches", outfilepath = "data")
}
}
