% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data based on input information}
\usage{
generate_data(data, obs = 1000)
}
\arguments{
\item{data}{Data input containing values for each of the associated parameters for multiple observations}

\item{obs}{(numeric) Number of samples to draw for bootstrapping procedure}
}
\value{
data base
}
\description{
Generate data based on input information
}
\examples{
\dontrun{
article_pdf_download(infilepath = "/data/isi_searches", outfilepath = "data")
}
}
