% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_data.R
\name{samp_data}
\alias{samp_data}
\title{Resample data}
\usage{
samp_data(data, n = 10000)
}
\arguments{
\item{data}{Data input containing values for each of the associated parameters for multiple observations}

\item{n}{(numeric) Number of samples to draw for bootstrapping procedure}
}
\value{
data base
}
\description{
Resample data
}
\examples{
\dontrun{
article_pdf_download(infilepath = "/data/isi_searches", outfilepath = "data")
}
}
